+++ lib/sidekiq/client.rb
@@ -6,10 +6,6 @@ require 'sidekiq/middleware/client/unique_jobs'
 module Sidekiq
   class Client
 
-    def self.middleware
-      raise "Sidekiq::Client.middleware is now Sidekiq.client_middleware"
-    end
-
     def self.default_middleware
       Middleware::Chain.new do |m|
         m.add Middleware::Client::UniqueJobs
@@ -24,22 +20,25 @@ module Sidekiq
       Sidekiq.redis { |x| x.smembers('queues') }
     end
 
+    # DEPRECATED
     def self.queue_mappings
       @queue_mappings ||= {}
     end
 
     # Example usage:
-    # Sidekiq::Client.push('my_queue', 'class' => MyWorker, 'args' => ['foo', 1, :bat => 'bar'])
-    def self.push(queue=nil, item)
-      raise(ArgumentError, "Message must be a Hash of the form: { 'class' => SomeClass, 'args' => ['bob', 1, :foo => 'bar'] }") unless item.is_a?(Hash)
+    # Sidekiq::Client.push('queue' => 'my_queue', 'class' => MyWorker, 'args' => ['foo', 1, :bat => 'bar'])
+    def self.push(item)
+      raise(ArgumentError, "Message must be a Hash of the form: { 'class' => SomeWorker, 'args' => ['bob', 1, :foo => 'bar'] }") unless item.is_a?(Hash)
       raise(ArgumentError, "Message must include a class and set of arguments: #{item.inspect}") if !item['class'] || !item['args']
+      raise(ArgumentError, "Message must include a Sidekiq::Worker class, not class name: #{item['class'].ancestors.inspect}") if !item['class'].is_a?(Class) || !item['class'].respond_to?('get_sidekiq_options')
 
-      queue = queue || queue_mappings[item['class'].to_s] || 'default'
-
-      item['class'] = item['class'].to_s if !item['class'].is_a?(String)
+      item['retry'] = !!item['class'].get_sidekiq_options['retry']
+      queue = item['queue'] || item['class'].get_sidekiq_options['queue'] || queue_mappings[item['class'].to_s] || 'default'
+      worker_class = item['class']
+      item['class'] = item['class'].to_s
 
       pushed = false
-      Sidekiq.client_middleware.invoke(item, queue) do
+      Sidekiq.client_middleware.invoke(worker_class, item, queue) do
         payload = MultiJson.encode(item)
         Sidekiq.redis do |conn|
           conn.multi do
@@ -52,16 +51,14 @@ module Sidekiq
       pushed
     end
 
-    # Please use .push if possible instead.
-    #
-    # Example usage:
+    # Redis compatibility helper.  Example usage:
     #
     #   Sidekiq::Client.enqueue(MyWorker, 'foo', 1, :bat => 'bar')
     #
     # Messages are enqueued to the 'default' queue.
     #
     def self.enqueue(klass, *args)
-      push(nil, { 'class' => klass.name, 'args' => args })
+      push('class' => klass, 'args' => args)
     end
   end
 end