+++ lib/sidekiq/middleware/client/unique_jobs.rb
@@ -7,23 +7,27 @@ module Sidekiq
       class UniqueJobs
         HASH_KEY_EXPIRATION = 30 * 60
 
-        def call(item, queue)
-          payload_hash = Digest::MD5.hexdigest(MultiJson.encode(item))
-          unique = false
+        def call(worker_class, item, queue)
+          enabled = worker_class.get_sidekiq_options['unique']
+          if enabled
+            payload_hash = Digest::MD5.hexdigest(MultiJson.encode(item))
+            unique = false
 
-          Sidekiq.redis do |conn|
-            conn.watch(payload_hash)
+            Sidekiq.redis do |conn|
+              conn.watch(payload_hash)
 
-            if conn.get(payload_hash)
-              conn.unwatch
-            else
-              unique = conn.multi do
-                conn.setex(payload_hash, HASH_KEY_EXPIRATION, 1)
+              if conn.get(payload_hash)
+                conn.unwatch
+              else
+                unique = conn.multi do
+                  conn.setex(payload_hash, HASH_KEY_EXPIRATION, 1)
+                end
               end
             end
+            yield if unique
+          else
+            yield
           end
-
-          yield if unique
         end
 
       end